requirements.txt
flask
pytest-docker-compose
detoxify
pytorch-lightning